scene: Replace speed with gpx
In order to pave the way for displaying other information from the gpx
file, the speed tag is renamed to gpx. The gpx object now has boolean
properties to choose whether to display the speed and elevation. The
speed keyword can be used as an alias for the gpx tag and it
automatically enables the speed.