scene: Store double timestamps instead of integer frame numbers
The key frames are now stored as a double and the script can specify
them as floating point values. Currently the script doesn’t know the
units of the timestamps and flootay treats them as frame numbers still.
In the future the idea is to make them always be a number of seconds.