flootay: Don’t add a final transparent frame
Previously a final transparent frame was added from the flootay output
so that the ffmpeg overlay filter wouldn’t copy the final drawing onto
the rest of the video if the input video is longer than the overlay.
This has the problem that it makes the video one frame longer and the
final frame never has any overlay. This is bad if the overlay is being
used to cover up something in the source video right until the last
frame because in that case it would always reveal the information for a
split second.

The original problem is no longer a problem now that speedy passes the
eof_action=pass to the overlay filter.