scene: Add a list of subobjects to the GPX object
Instead of having a bool for each possible thing to display from the GPX
data, there is now a list of them in each GPX object. That will
eventually allow adding properties to the GPX objects and it also means
the user can control the order they are displayed in.