speedy: Include the gpx offset of all videos
Instead of only looking at videos that have been mentioned in the
script, it now looks at the length of all video files in the current
directory when working out the GPX offsets. That way it’s not necessary
to use all of the videos that have an offset specified in the script and
it’s possible to start from the second part of a set of a videos when
the offset is only specified on the first part.